requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.12.0
edx-opaque-keys[django]

# Development dependencies
black>=22.0.0
isort>=5.10.0
flake8>=4.0.0
pytest>=7.0.0
pytest-django>=4.5.0
coverage>=6.0.0